+++ src/components/AboutContainer/AboutContentContainer/AboutContentPages/index.js
@@ -1,0 +1,5 @@
+import Programs from './Programs';
+import WhatWeDo from './WhatWeDo';
+import WhoWeAre from './WhoWeAre';
+
+export { Programs, WhatWeDo, WhoWeAre };
